Changed the following:
However, we believe that these details about the data mining aspect of our
model are important enough that they belong in the main text of the
article. We have attempted to add more intuitive explanation of the technical  approaches so that readers can potentially skip the technical details about  model definition and estimation.